+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The minor version will be incremented upon a breaking change and the patch versi
 ### Fixes
 
 - richat: fix `ping_interval` typo ([#148](https://github.com/lamports-dev/richat/pull/148))
+- richat: fix `x_tokens` typo ([#149](https://github.com/lamports-dev/richat/pull/149))
 
 ### Features